requirements.txt
# Telegram Bot Framework
aiogram>=3.25.0

# Video/Audio Downloader
yt-dlp>=2026.02.21

# Redis (cache + task queue)
redis[hiredis]>=5.0.0

# Async Task Queue
arq>=0.26.1

# Fast Event Loop (Linux only, fallback on Windows)
uvloop>=0.21.0; sys_platform != 'win32'

# Async File Operations
aiofiles>=23.2.1

# Environment Variables
python-dotenv>=1.0.0

# HTTP client (already in aiogram deps, but explicit)
aiohttp>=3.10.0

# Shazam — musiqa aniqlash
shazamio>=0.7.0

# System monitoring (admin panel uchun)
psutil>=6.0.0
